ChangeRebuilderImpl: Ensure first event creates change
There were some corner cases where the first event according to
EVENT_ORDER might not be a PatchSetEvent, or might not even be by the
change owner at all. This breaks ChangeNotesParser, which blindly
takes as the change owner the author of the first commit in the notes
graph.

One such corner case is where a historic comment is added prior to
change creation, which gets truncated to have the same timestamp as
change creation, and inserted first in the list. In this case it
should simply come after the corresponding PatchSetEvent.

Another corner case is more subtle: races can cause a historic comment
to get inserted between the createdOn timestamp of the change and the
createdOn timestamp of the first patch set (or first remaining patch
set). This would not cause the timestamp to get truncated, but would
predate any PatchSetEvent, so we can't depend on the tiebreaker used
in the previous corner case. As a fallback, double-check that the very
first event in the list, after sorting, includes the author and other
required fields. It ends up looking a little odd that this commit has
a patch set ID whose revision has not been parsed yet, but
ChangeNotesParser can handle this.

Change-Id: If8ef07268126c19b842cb4b226dc284c6fd4f2b4

Former-commit-id: 1d96c9214a38608e5038e202845d9de73a147a3b