Support fetch of refs/users/self from All-Users repository
Allow a user to fetch its refs/users/YY/XXXXXXX branch as
refs/users/self. The user branch contains the sharded account ID in
the ref name, but this is an implementation detail that should be
hidden from the user. On the other hand users should be able to access
their user branch easily.

Change-Id: I0f5a9f1456a86873ff91ad418aa60b1eee568109
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: d3eb8fc937569599843cda0a7024d99402aee5f8