BatchUpdate: Bump lastUpdatedOn by default
There are only two cases where we don't currently bump lastUpdatedOn:
adding reviewers and adding draft comments. The vast majority of other
cases we do want to bump lastUpdatedOn.

Change the ChangeContext#saveChange method to #bumpLastUpdatedOn, to
be explicit about what this actually does. Default to true to make it
easier on callers.

Change-Id: I3d8b13f70b082f8f4ef15df02f72840ed42eed85

Former-commit-id: 281dc411eee5a7e042d9a282ac70dfbdff19b131