Fix migration of SSH keys with invalid sequence numbers (<= 0)
Sequence numbers for SSH keys start at 1, however the InitAdminUser
init step creates an initial SSH key with sequence number 0. Having a
key with sequence number 0 breaks the migration to schema 124 and lets
it fails with ArrayIndexOutOfBoundsException.

Fix any key with sequence number <= 0 during the migration and use the
next unused valid sequence number for it.

Change-Id: I772cc9b0fc71aea7fb97014d0e92cbd99d1d2fe7
Signed-off-by: Edwin Kempin <ekempin@google.com>

Former-commit-id: 2d7189549b35a7754b17c58e6756fb177cf5cf75