NoSuchChangeException not thrown by ChangeNotes with non-existing id
createChecked relied on a ChangeNotes#getChange() to return null to
identify whether or not a change existed. This did not work as expected
since the change cannot be null, instead another exception was thrown.

Now, all createChecked alternatives check if the change exists so that
an exception is thrown before trying to create the ChangeNote with a
null change.

Change-Id: I02d925ad0b0fa9d9ac995d37085c119eff47ca9c

Former-commit-id: c1dc36d135a077f3bdd9d87cd4c3bc8cc1f97044