ChangeBundle: Ignore originalSubject between ReviewDb and NoteDb
The migration path in Change#setCurrentPatchSet means that the
original subject field in ReviewDb may be null, or may be the actual
original subject of PS1, or may be the subject of any intervening
patch set, if that patch set was the first one pushed after adding the
column to ReviewDb. This means we can't reliably compare these after
converting to NoteDb, at least without changing the implementation of
ChangeBundle substantially in order to give it access to the
repository.

Satisfy ourselves with explicitly checking the subject and
originalSubject in ChangeRebuilderIT to ensure it's populated properly
in the common case.

Change-Id: I2160bf6f38b5ea97e0c3ee512a0450854e1af265

Former-commit-id: 7672de4d265d65924d82cb94e1e763559aebf5ce