ChangeBundle: Rewrite ChangeMessage comparison again
The previous implementation attempted to stably sort both lists and do
pairwise comparison. Unfortunately the sorting was flawed: it sorted
by timestamp first, and then by other fields. When change timestamps
get rounded in NoteDb, this means that the order might get swapped
when we fall back to comparing on the next field.

Switch to a more naive algorithm: iterate through both lists of
messages and remove matching pairs (using the same fuzzy equivalence
we have been using previously). The sets are equivalent only if we end
up with two empty lists. This is obviously quadratic, but we can
improve it slightly by limiting our search of the B list to only
ChangeMessages where the author/message/tag definitely already match.

Change-Id: I23f52531deb446c9f8c6acec33efd1a8f35a1336

Former-commit-id: 0c1ac02179e0e82f2e0e32abf4bb0fa07f21cba4