ChangeBundle: Ignore lastUpdatedOn if bundle has no entities
I found a few changes in our dataset where the lastUpdatedOn timestamp
of the Change was later than any other entity, but no other fields
differed so no FinalUpdatesEvent was added to the list. This is in
theory possible, for example if the change state changed multiple
times but ended up in the original state. There's not much we can do
on the NoteDb side if we don't want to add a dummy event just to
update the timestamp, which we don't, really.

Just allow NoteDb lastUpdatedOn timestamps to match either the
timestamp on the Change, or the latest timestamp among all entities
other than the Change itself.

Change-Id: I3d933b878747fae76c5509107e6cedbd9cd5115a

Former-commit-id: e46af3e2b876008c0a54d502ab527c8906a09e74