NoteDb: Don't convert entities whose patch set is missing
The most obvious case where one might expect to find entities (e.g.
comments) whose corresponding patch set is missing is when a draft
patch set is deleted, but in practice that shouldn't happen, since
those entities are supposed to be deleted along with the patch set.

In fact though, there is at least one other weird case where we can
have entities for missing patch sets: somehow we ended up with some
comments on PS0[1], which of course doesn't exist. (This number is
used internally for change edits, but it should be impossible to add
an inline comment on an edit.)

Explicitly filter these items out of NoteDb, since the parser can get
confused if a commit refers to a patch set ID that isn't eventually
filled in. They were already implicitly filtered out in the case of
PatchLineComments, which is how I detected the cited example.

In ChangeBundle, limit the scope of all comparisons to entities whose
corresponding patch sets also exist in the bundle. In the case where
one bundle has extra patch sets and also other extra entities
associated with those patch sets, this means we will only get a diff
message for the mismatched patch sets. That's ok, it still flags it as
an issue.

[1] One example: https://gerrit-review.googlesource.com/69777

Change-Id: I0f9cdbf347acc6f478d5483967d275bd01591b9b

Former-commit-id: 071e32506ba98701a4fd33b7b9aca7e10302fe34