Add a short-lived ref cache interface
We already have ChainedReceiveCommands that contains a per-instance
ref cache that guarantees that subsequent reads for a given ref will
be the same, and avoids a few syscalls to boot. Extract an interface
from ChainedReceiveCommands so we can use the caching part without the
commands part.

Use Optional<ObjectId> consistently in the return type, so that
callers do not have to determine whether null or zeroId indicates a
missing ref. Currently they need to expect null for exactRef-style
semantics and zeroId for ReceiveCommand-style semantics, which can be
confusing.

Change-Id: I5e02032d6d406b30a1cc4704808e9b2475e5b417

Former-commit-id: a4d86e8f40b22498ac6217ab7b9a6b5f2e4b43e1