NoteDbUpdateManager: Refuse to update out-of-date NoteDb state
If a ReviewDb write succeeds but the corresponding NoteDb write fails,
then the NoteDbChangeState in ReviewDb will be out of sync with what's
in ReviewDb. This is what's supposed to happen, and typically a change
will get rebuilt on the next time it is read so the NoteDbChangeState
gets back in sync.

However, there is still a race condition by which a change might be
stale and still not get auto-rebuilt before applying a NoteDb update:

 1. Thread A reads ChangeNotes in order to pass to a ChangeUpdate, and
    finds that the change is up to date.
 2. Thread B reads ChangeNotes, then creates and executes a
    ChangeUpdate, but the NoteDb write fails.
 3. Thread A stages its NoteDb update and writes the resulting
    NoteDbChangeState into ReviewDb.
 4. Thread A applies its ChangeUpdate to NoteDb.

This is a bad outcome: now the NoteDbChangeState is considered up to
date even though the meta graph is completely missing the write from
thread B.

Add a bit more defense in depth to avoid this problem, and have
NoteDbUpdateManager ensure that the state is up-to-date before
attempting to write, throwing OrmConcurrencyException if not. This
uses the same ChainedReceiveCommands as is used for actually updating
the ref, so even if it happens to lose another race when reading the
ref, that would result in LOCK_FAILURE when actually trying to update
NoteDb.

After this change, Thread A will leave the NoteDbChangeState in the
same state that B left it after (2), but otherwise proceed with its
ReviewDb write. This means the state is out of sync with NoteDb, which
is what we want, and should trigger a rebuild on the next read.

Change-Id: I779a0e041f3b96d4eabac6bd675ea9a4e33c6cec

Former-commit-id: edabbdaa68891ee5a9a8430df19e9edef73a0cc5