Always use a thread pool for interactive indexing
The problem is that when you cancel an ssh command, the thread handling
the command gets interrupted. If it happens that the thread is
interrupted while it's updating the index, this will close the index
writer and any subsequent operations that need indexing will fail.

In order to protect Lucene index writer from any possible thread
interruption, we need to delegate the index writing to a separate thread
pool and never cancel or interrupt threads from that thread pool.

For background indexing, Gerrit already uses a thread pool but for
interactive indexing, Gerrit optionally support using a thread pool and
the default configuration is not using one.

Remove the possibility to do interactive indexing in the same thread as
the operation to prevent Lucene index writer from being interrupted.

Bug: Issue 3885
Change-Id: Ic20cd1caa827a205a83417eff7a619133bc2450c

Former-commit-id: 176cd58bb0bb527e865fd63f68f4d9ef277ab76f