ChangeRebuilderImpl: Skip patch sets greater than current
AOSP has some changes[1] dating from before we used transactions
properly, where there is a PatchSet entity with an ID higher than the
currentPatchSetId in the Change entity. In NoteDb we don't record the
current patch set ID explicitly, we just use the greatest
(non-deleted) ID found in the meta graph, so we can't represent this
discrepancy.

Just don't convert these newer patch sets or their associated
entities, and teach ChangeBundle to ignore this diff. This means we
might end up discarding some comments of the form "What's the deal
with this extra patch set?", which seems acceptable.

[1] https://android-review.googlesource.com/31041

Change-Id: I30e55f6455fe8db3dac3d2502e5e5eafc5bc2438

Former-commit-id: f3a0051023875fa6de495072f0cf6b106e83b559