ChangeBundle: Always filter out patch sets greater than current
The logic in I30e55f6455 for ignoring diffs between entities whose
patch set ID exceeds the current patch set ID of the change was
flawed. It wasn't excluding entities other than the patch set itself.

Fix this by filtering out all entities for patch sets that are not
valid, i.e. don't have corresponding PatchSet entities or whose ID
exceeds the current ID on the change. This keeps the logic simpler, at
the cost of not detecting this difference when two ReviewDb bundles
are compared, which is not a huge loss as the class is mainly for
comparing ReviewDb to NoteDb.

Change-Id: I2da1945c9b2a07b54b0c53185c6d5acacfc706cb

Former-commit-id: a5778de8e126eac03bc823553f8d9ffdc919f8ea