Factor out a value class for the output of ChangeNotesParser
The intent of this value class is to be in-process cacheable, so we
don't have to repeatedly read objects from the repo to parse and
re-parse the notes graph. Thus it contains the full set of Change
fields and associated Change entities.

Note that, even though it's somewhat unwieldy, we explicitly list
each field in Change that ChangeNotesParser knows how to parse, and
do *not* store a Change in ChangeNotesState. This is because the
rowVersion and noteDbState fields are used internally to test
consistency between ReviewDb and NoteDb, so we can't pretend that
these were populated from NoteDb and store them in any sort of cache.
Instead, we store the other fields, and provide a single method to
copy the field values to a Change object that we know was read from
ReviewDb. This is exactly what already happens in the
ChangeNotes.Factory path.

The main output of ChangeNotesParser is a ChangeNotesState, and each
ChangeNotes has a single one of these from which it can serve most of
its public methods. These methods were intentionally already
returning immutable collections in anticipation of this day.

Change-Id: I572cb1a8cb9ce8f68c2e4b744bdfab06a3383a1e

Former-commit-id: f3fe1a0c270ab349d471b545e81d28c5c729ef6d