Cache parsed ChangeNotes
By instrumenting a dev server, I found that loading the change screen
for a single change plus 2 related changes and 2 conflicting changes
results in a whopping 25 loads of a ChangeNotes. On a SQL database
that's not too terrible, but git storage is enough slower and more
memory-intensive than SQL that this is noticeable. Plus, the
ChangeNotes approach is to parse EVERYTHING, which is often more than
we need.

Solve this problem in the way we solve all problems: throw a cache in
front of it. This data has the nice property that the entities parsed
from a particular SHA-1 are immutable forever, so we don't need to
worry about invalidation like we would with a cache keyed by, say,
change number.

Change-Id: I4bbb9bf82f8f228b3ae8333a288af35d163c050f

Former-commit-id: 6f8254b6671af004d633c5120d3c1787294b3124