RelatedChanges: Show "Submitted Together" also for changes with topic
The reason for hiding the Submitted Together tab is that when
SubmitWholeTopic option is enabled, the same changes would appear twice
- in Submitted Together and in Same Topic.

But when some changes have topic, and the Submit Whole Topic option is
disabled, there's no reason to hide the Submitted Together tab. This tab
also shows changes that are submitted together because of project submit
strategy, like Merge or Rebase If Necessary.

Change-Id: Ibbafbdbea921444278efc602801d28b756145f25

Former-commit-id: 7dbaf5706b81afdf9e5e30fec51b1c86d1522a5b