Synchronize MyersDiff and HistogramDiff invocations on local variable
I4516bcc2c synchronized MyersDiff and HistogramDiff invocations but it
used a synchronization object that is not local to the PatchList being
loaded. In other words, even if PatchList can be loaded in parallel,
only one file header diff can be computed at the time across the entire
server.

In stable-2.12, a fresh instance of PatchListLoaded is used every time a
PatchList is loaded so this change is not needed, and must be reverted
when merging stable-2.11 into stable-2.12.

Change-Id: I5acace093b16c9b87cf465c515018433c3f38d6a

Former-commit-id: 776e004f059d4d4b4d6e3f24490aa0336762f614