SubmoduleOp: Drop ReviewDb from parameter list
Since I8178e0150ef1f708e3026db8858846498c3adb9a SubmoduleOp did not need
to pass around the reviewDb as the MergeOpRepoManager should handle all
the queries to the database.

It was not detected however as updateSuperProjects and updateGitlinks
called each other recursively.

As the recursion is gone in I5cd1d719e77dae1db2e359270256fb9a25adce02,
it should have been detected and fixed in that change.

Change-Id: I99f2f1b5ba7cb75dc6a0b194ce58f8ed69a3faa7
Signed-off-by: Stefan Beller <sbeller@google.com>

Former-commit-id: 52977ed58d9bbaef41be088f4de905b626348d53