Allows topic editing in the change view
Adds an editable text label to the gr-change-metadata for modifying the
topic of a change.  Introduces the gr-editable-label as a UI element for
editable text labels.  Editing the topic is only allowed if the user is
logged in and has the topic-edit permission.

This is a redo of change 78110, which had been rebased wrong and had to
be reverted. https://gerrit-review.googlesource.com/c/78110/

Bug: Issue 4102
Change-Id: Ia7bb7b803a9fdd15a5f9f0f8eecf77bdd3ee4ba7

Former-commit-id: f1d7d87d6fa4fcea66451599a3990b4ac844bf29