ChangeBundle: Fix ChangeMessage comparison for extra patch sets
We were only excluding messages for patch sets that don't exist in the
full, unfiltered set of PatchSet.Ids. But there might be extraneous
patch sets that actually do exist in the bundle, but are greater than
the current patch set ID on the change. We were already filtering
these out for other entity types in limitToValidPatchSets, just not
for ChangeBundle, which uses a slightly different implementation.

Change-Id: Id1a0a9462f4096deee7a1d1a6eb673bc37529958

Former-commit-id: cb4c481f77d5afe714ee4256d44e91f03e08208c