Revert "Allows topic editing in the change view
Adds an editable text label to the gr-change-metadata for modifying the
topic of a change.  Introduces the gr-editable-label as a UI element for
editable text labels.  Editing the topic is only allowed if the user is
logged in and has the topic-edit permission.

Bug: Issue 4102
Change-Id: I42e8f65bd538ec50f12d29ef886395b6a70edbb7
"

This reverts commit fa6bdf7a0a387f35db9fc712005e58e096000141 [formerly 9ffd7b57a3835593c0f58a2c4d383d7619c67243].

Reason for revert: Not the correct change to submit. A lot of stuff was undone.

Change-Id: Ie51d712f2cba87a5e27639381f7f8126cc024833

Former-commit-id: 445f549f9c147841582cda58e8f6012e185f3d7a